requirements.txt
python-telegram-bot
requests
beautifulsoup4
psycopg2-binary